Merge program-manager branch with main.

+++ src/client/shaders/color/fragment.glsl
@@ -1,9 +1,7 @@
 #version 300 es
-
-precision lowp float;
+precision highp float;
 
 in vec4 vColor;
-
 out vec4 fragColor;
 
 void main() {
